Removed ConcurrentLogEventSink for sake for future enhancements
