+++ src/Serilog.Sinks.AzureDocumentDb/Sinks/AzureDocumentDb/AzureDocumentDbSink.cs
@@ -13,16 +13,20 @@
 // limitations under the License.
 
 using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Azure.Documents;
 using Microsoft.Azure.Documents.Client;
 using Serilog.Core;
+using Serilog.Debugging;
 using Serilog.Events;
 
 namespace Serilog.Sinks.AzureDocumentDb
 {
-    public class AzureDocumentDBSink : ConcurrentLogEventSink
+    public class AzureDocumentDBSink : ILogEventSink, IDisposable
     {
         readonly IFormatProvider _formatProvider;
         DocumentClient _client;
@@ -39,6 +43,8 @@ namespace Serilog.Sinks.AzureDocumentDb
 
             CreateDatabaseIfNotExistsAsync(databaseName).Wait();
             CreateCollectionIfNotExistsAsync(collectionName).Wait();
+
+            InitializeParallelSink();
         }
 
         private async Task CreateDatabaseIfNotExistsAsync(string databaseName)
@@ -62,7 +68,13 @@ namespace Serilog.Sinks.AzureDocumentDb
             }
         }
 
-        protected override void WriteLogEvent(LogEvent logEvent)
+        #region Parallel Log Processing Support
+        private CancellationTokenSource _cancelToken = new CancellationTokenSource();
+        private BlockingCollection<LogEvent> _logEventsQueue;
+        private Thread _workerThread;
+        private List<Task> _workerTasks = new List<Task>();
+
+        private void WriteLogEvent(LogEvent logEvent)
         {
             _client.CreateDocumentAsync(
                 _collection.SelfLink,
@@ -72,5 +84,78 @@ namespace Serilog.Sinks.AzureDocumentDb
                     _storeTimestampInUtc),
                 new RequestOptions { }, false).Wait();
         }
+
+        void InitializeParallelSink()
+        {
+            _logEventsQueue = new BlockingCollection<LogEvent>(1000);
+            _workerThread = new Thread(Pump) { IsBackground = true, Priority = ThreadPriority.AboveNormal };
+            _workerThread.Start();
+        }
+
+        void Pump()
+        {
+            try
+            {
+                var numProcessors = Environment.ProcessorCount;
+                while (true)
+                {
+                    var next = _logEventsQueue.Take(_cancelToken.Token);
+                    var workerTask = Task.Factory.StartNew((t) =>
+                    {
+                        WriteLogEvent(t as LogEvent);
+                    }, next);
+
+                    _workerTasks.Add(workerTask);
+                    if (_workerTasks.Count >= numProcessors)
+                    {
+                        Task.WaitAll(_workerTasks.ToArray());
+                        _workerTasks.Clear();
+                    }
+                }
+            }
+            catch (OperationCanceledException)
+            {
+                Task.WaitAll(_workerTasks.ToArray());
+                _logEventsQueue.AsParallel().ForAll(item => WriteLogEvent(item));
+            }
+            catch (Exception ex)
+            {
+                SelfLog.WriteLine("{0} fatal error in worker thread: {1}", typeof(AzureDocumentDBSink), ex);
+            }
+        }
+
+        #endregion
+
+        #region IDisposable Support
+        private bool disposedValue = false;
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (!disposedValue)
+            {
+                if (disposing)
+                {
+                    _cancelToken.Cancel();
+                    _workerThread.Join();
+                }
+
+                disposedValue = true;
+            }
+        }
+
+        public void Dispose()
+        {
+            Dispose(true);
+        }
+
+        #endregion
+
+        #region ILogEventSink Support
+        public void Emit(LogEvent logEvent)
+        {
+            _logEventsQueue.Add(logEvent);
+        }
+
+        #endregion
     }
 }
